Move compute_residue_frames to data_processing
