+++ requirements/requirements.process.txt
@@ -1,3 +1,4 @@
 biopython
+numpy
 pyspark
 pytest
